+++ lib/browser/browser_interface.ts
@@ -1,0 +1,6 @@
+import Browser from './browser';
+import BrowserAssertions from './browser_assertions';
+
+export default interface BrowserInterface extends Browser {
+    readonly assert: BrowserAssertions;
+}