+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * Wendigo refactored into TypeScript
+* Changes in plugin system. Dropped support for not assertions
 * Removed deprecated methods assert.cookie, assert.webworker and assert.not.cookie
 * Browser.request.all changed into a function to make it consistent with other modules
 * WaitForRequest/Response and waitForNextRequest/Response moved to requests module